File System and capabilities are working as I would expect them. Need to add UI in still but the scaffolding is here
