+++ src/learning_mode/start_plm_logging.ps1
@@ -1,5 +1,2 @@
-#TODO replace with xperf
-
-$AcpPath = "c:\users\adminuser\desktop\acp"
-import-module "$AcpPath\Microsoft.Windows.Win32Isolation.ApplicationCapabilityProfiler.dll"
-start-profiling -force
+$wprp = Join-Path $PSScriptRoot 'PLM.wprp'
+wpr -start "$wprp!AccessFailureProfile" -filemode
